Move container to footer
