+++ dashboard/src/containers/FooterContainer/index.ts
@@ -1,0 +1,3 @@
+import FooterContainer from "./FooterContainer";
+
+export default FooterContainer;
