--- dashboard/src/containers/LayoutContainer/index.ts
@@ -1,3 +1,0 @@
-import LayoutContainer from "./LayoutContainer";
-
-export default LayoutContainer;
