+++ dashboard/src/containers/Root.tsx
@@ -2,10 +2,10 @@ import { ConnectedRouter } from "connected-react-router";
 import * as React from "react";
 import { Provider } from "react-redux";
 
+import Layout from "../components/Layout";
 import store, { history } from "../store";
 import ConfigLoaderContainer from "./ConfigLoaderContainer";
 import HeaderContainer from "./HeaderContainer";
-import Layout from "./LayoutContainer";
 import Routes from "./RoutesContainer";
 
 class Root extends React.Component {
